+++ src/model/app/reducers.js
@@ -46,6 +46,8 @@ import {
   SET_REALTIME_KEY,
   SET_DISPLAY_MENU,
   SET_STOPS_URL,
+  SET_MAX_CANVAS_SIZE,
+  SET_EXPORT_PRINT_OPTIONS,
   SET_GEOLOCATING,
   SET_FOLLOWING,
   SET_OVERLAY_ELEMENT,
@@ -53,6 +55,7 @@ import {
 
 import SearchService from "../../components/Search/SearchService";
 import { isOpenedByMapset } from "../../utils/redirectHelper";
+import { LS_SIZE_KEY } from "../../utils/constants";
 
 const dftlInteractions = defaultInteractions({
   altShiftDragRotate: false,
@@ -102,6 +105,10 @@ const getInitialState = () => ({
   consentGiven: false,
   disableCookies: false,
   displayMenu: true,
+  maxCanvasSize:
+    localStorage.getItem(LS_SIZE_KEY) &&
+    parseFloat(localStorage.getItem(LS_SIZE_KEY)),
+  exportPrintOptions: null,
   isGeolocating: false,
   isFollowing: false,
 });
@@ -334,6 +341,16 @@ export default function app(state = getInitialState(), action) {
       return {
         ...state,
         stopsUrl: action.data,
+      };
+    case SET_MAX_CANVAS_SIZE:
+      return {
+        ...state,
+        maxCanvasSize: action.data,
+      };
+    case SET_EXPORT_PRINT_OPTIONS:
+      return {
+        ...state,
+        exportPrintOptions: action.data,
       };
     case SET_OVERLAY_ELEMENT:
       return {